+++ portal-impl/src/com/liferay/portal/jsonwebservice/SpringJSONWebServiceScannerStrategy.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.jsonwebservice.scanner;
+package com.liferay.portal.jsonwebservice;
 
 import com.liferay.portal.kernel.jsonwebservice.JSONWebServiceScannerStrategy;
 import com.liferay.portal.kernel.util.ProxyUtil;
